+++ godog-hello/greeter/greeter.go
@@ -1,4 +1,4 @@
-package main
+package greeter
 
 type Greeter struct{}
